feat(nvim): set filetype highlighting with treesitter
A severe limitation is the headers at the top of these files, as I don't
know how to write treesitter injections at this point in time to give
the headers highlights separate from the body. Another issue is that
this forces the users to opt-in to seeing headers which may not be
desired. But for a PoC? Good enough.

Related to #7